A-HIGH-NNN.add_composite_tree_validations
