+++ docs/feature_stories/FS_00_13_77_97.plugin_framework.md
@@ -8,12 +8,13 @@ feature_status: TEST
 
 Plugins are configured under `plugin_instance_entries` in `argrelay_plugin.yaml` config file (see also `PluginConfigSchema.py`).
 
-Both plugin and plugin instance are used interchangeably to mean individual (runtime) plugin instance
+There are plugin classes (implementations),
+but the framework mostly deals with individual (runtime) plugin instances
 identified by `plugin_instance_id` and configured under `plugin_instance_entries`.
 
-Each plugin in configured in one place, but do not confuse (distinguish):
+Each plugin instance is configured in one place, with two layers
 *   `plugin_entry` which `argrelay` uses to manage plugin instances.
-*   `plugin_config` (part of `plugin_entry`) which is plugin-specific config (opaque to `argrelay`).
+*   `plugin_config` (part of `plugin_entry`) which is plugin-specific config (opaque to `argrelay` layer).
 
 Each plugin entry under `plugin_instance_entries` is defined by `PluginEntrySchema.py`.
