+++ docs/feature_stories/FS_33_76_82_84.composite_tree.md
@@ -5,7 +5,9 @@ feature_status: TEST
 ---
 Keywords: `composite`, `tree`, `composite_tree`
 
-TODO_26_08_72_06: interp vs delegator: consider making it clear how FT-interp is different (and required to exists) compared to delegator
+TODO: TODO_09_32_52_34: rename `composite_tree` to `composite_forest`
+
+TODO: TODO_26_08_72_06: interp vs delegator: consider making it clear how FT-interp is different (and required to exists) compared to delegator
 
 TODO: This feature will (potentially) combine and replace (or at least affect):
 *   FS_01_89_09_24 interp tree (except the FS_42_76_93_51 very first zero arg use case).
