+++ docs/release_notes/v0.6.1.final.md
@@ -4,4 +4,4 @@
 *   Improve spinner - restore command line chars overwritten by spinner (Bash fails to do it in all cases).
 *   Disable debug output in frequently used scripts like `@/exe/dev_shell.bash`.
 *   Make `@/exe/init_shell_evn.bash` report argrelay version, conf dst, venv path, server URL.
-*   Upgrade `DefaultConfigurator` to provide GUI with details on: project title, project page URL, conf dir URL.
+*   Upgrade `ConfiguratorDefault` to provide GUI with details on: project title, project page URL, conf dir URL.
