+++ docs/release_notes/v0.6.2.final.md
@@ -1,4 +1,4 @@
 
 2024 JAN 31
 
-*   Fix `DefaultConfigurator` to use `realpath` to derive inclusion relationship between paths.
+*   Fix `ConfiguratorDefault` to use `realpath` to derive inclusion relationship between paths.
