+++ docs/task_refs/TODO_09_32_52_34.rename_to_composite_forest.md
@@ -1,0 +1,6 @@
+
+TODO: TODO_09_32_52_34: rename `composite_tree` to `composite_forest`
+
+We mostly deal with independent roots, so it is more appropriate to call it forest.
+
+Even if we combine them, they are not combined via common node, but via a separate `CompositeForestSchema`.
