+++ docs/task_refs/TODO_99_87_25_42.next_incompatible_changes.md
@@ -22,9 +22,9 @@ TODO:
     *   `propose_arg_values` -> `complete_line`
     *   `relay_line_args` -> `invoke_line`
 
-*   TODO: Rename `DefaultConfigurator` to `GuiConfigurator`.
-*   TODO: Move all GUI config into DefaultConfigurator (including banner).
-*   TODO: Move `gui_banner_config` under `DefaultConfigurator`
+*   TODO: Rename `ConfiguratorDefault` to `GuiConfigurator`.
+*   TODO: Move all GUI config into ConfiguratorDefault (including banner).
+*   TODO: Move `gui_banner_config` under `ConfiguratorDefault`
 
 *   TODO: Rename all `func_id` to be prefixed with `func_id`, for example, `func_id_query_enum_items` -> `func_id_query_enum_items`.
 *   TODO: Rename: prefix all `func_id` with `func_id`, for example: `goto_service_func` -> `func_id_goto_service`.
