+++ src/argrelay/plugin_config/ConfiguratorDefault.py
@@ -9,9 +9,9 @@ from argrelay.custom_integ.git_utils import (
     get_git_repo_root_path,
 )
 from argrelay.misc_helper_common import get_argrelay_dir
-from argrelay.plugin_config.AbstractConfigurator import AbstractConfigurator
-from argrelay.plugin_config.DefaultConfiguratorConfigSchema import (
-    default_configurator_config_desc,
+from argrelay.plugin_config.ConfiguratorConsistent import ConfiguratorConsistent
+from argrelay.plugin_config.ConfiguratorDefaultConfigSchema import (
+    configurator_default_config_desc,
     commit_id_url_prefix_,
     git_files_by_commit_id_url_prefix_,
     project_title_,
@@ -19,15 +19,15 @@ from argrelay.plugin_config.DefaultConfiguratorConfigSchema import (
 )
 
 
-class DefaultConfigurator(AbstractConfigurator):
+class ConfiguratorDefault(ConfiguratorConsistent):
 
     def load_config(
         self,
         plugin_config_dict,
     ) -> dict:
-        return default_configurator_config_desc.dict_from_input_dict(plugin_config_dict)
+        return configurator_default_config_desc.dict_from_input_dict(plugin_config_dict)
 
-    def provide_project_title(
+    def _provide_project_title(
         self,
     ) -> Union[str, None]:
         if project_title_ in self.plugin_config_dict:
@@ -35,7 +35,7 @@ class DefaultConfigurator(AbstractConfigurator):
         else:
             return None
 
-    def provide_project_page_url(
+    def _provide_project_page_url(
         self,
     ) -> Union[str, None]:
         if project_page_url_ in self.plugin_config_dict:
@@ -43,7 +43,7 @@ class DefaultConfigurator(AbstractConfigurator):
         else:
             return None
 
-    def provide_project_git_files_by_commit_id_url_prefix(
+    def _provide_project_git_files_by_commit_id_url_prefix(
         self,
     ) -> Union[str, None]:
         if git_files_by_commit_id_url_prefix_ in self.plugin_config_dict:
@@ -51,7 +51,7 @@ class DefaultConfigurator(AbstractConfigurator):
         else:
             return None
 
-    def provide_project_commit_id_url_prefix(
+    def _provide_project_commit_id_url_prefix(
         self,
     ) -> Union[str, None]:
         if commit_id_url_prefix_ in self.plugin_config_dict:
@@ -59,7 +59,7 @@ class DefaultConfigurator(AbstractConfigurator):
         else:
             return None
 
-    def provide_project_git_commit_id(
+    def _provide_project_git_commit_id(
         self,
     ) -> Union[str, None]:
         argrelay_dir = get_argrelay_dir()
@@ -68,7 +68,7 @@ class DefaultConfigurator(AbstractConfigurator):
         else:
             return None
 
-    def provide_project_git_commit_display_string(
+    def _provide_project_git_commit_display_string(
         self,
     ) -> Union[str, None]:
         argrelay_dir = get_argrelay_dir()
@@ -77,7 +77,7 @@ class DefaultConfigurator(AbstractConfigurator):
         else:
             return None
 
-    def provide_project_git_commit_time(
+    def _provide_project_git_commit_time(
         self,
     ) -> Union[int, None]:
         argrelay_dir = get_argrelay_dir()
@@ -86,7 +86,7 @@ class DefaultConfigurator(AbstractConfigurator):
         else:
             return None
 
-    def provide_project_git_repo_relative_argrelay_dir(
+    def _provide_project_git_repo_relative_argrelay_dir(
         self,
     ) -> Union[str, None]:
         argrelay_dir_abs_path = os.path.realpath(os.path.abspath(get_argrelay_dir()))
@@ -102,7 +102,7 @@ class DefaultConfigurator(AbstractConfigurator):
             # Wrap into slashes `/` for concatenation:
             return f"/{argrelay_dir_rel_path}/"
 
-    def provide_project_current_config_path(
+    def _provide_project_current_config_path(
         self,
     ) -> Union[str, None]:
         argrelay_dir = get_argrelay_dir()
