+++ src/argrelay/plugin_config/ConfiguratorDefaultConfig.py
@@ -4,9 +4,9 @@ from dataclasses import dataclass, field
 
 
 @dataclass
-class DefaultConfiguratorConfig:
+class ConfiguratorDefaultConfig:
     """
-    This class configures `DefaultConfigurator`.
+    This class configures `ConfiguratorDefault`.
     """
 
     project_title: str = field(default = "")
@@ -23,12 +23,12 @@ class DefaultConfiguratorConfig:
     """
     Provides a URL prefix to access files by commit id.
 
-    See also `AbstractConfigurator.provide_project_git_files_by_commit_id_url_prefix`.
+    See also `ConfiguratorAbstract.provide_project_git_files_by_commit_id_url_prefix`.
     """
 
     commit_id_url_prefix: str = field(default = "")
     """
     Provides a URL prefix to access page with commit id.
 
-    See also `AbstractConfigurator.provide_project_commit_id_url_prefix`.
+    See also `ConfiguratorAbstract.provide_project_commit_id_url_prefix`.
     """
