+++ tests/offline_tests/plugin_config/__init__.py
@@ -1,0 +1,3 @@
+"""
+"""
+# Intentionally empty: enables test discovery within this dir.
